% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyspan.R
\name{xyspan}
\alias{xyspan}
\title{Extract X-axis and Y-axis span}
\usage{
xyspan(eval_period, smooth_fd_object)
}
\arguments{
\item{eval_period}{A vector that represents the range of values at which the functional data object is to be evaluated against}

\item{smooth_fd_object}{A functional data object that contains the set(s) of smooth curves}
}
\value{
A data frame object that contains the corresponding integer value of the set of basis function coefficients used (\code{coef_set}),
the vector range at which the functional data object was evaluated (\code{evalarg}), the X-axis span (\code{x_span_vec}) and the Y-axis span of the each phase-plane plot (\code{y_span_vec}).
}
\description{
Generates a phase-plane plot and extracts the X-axis and Y-axis span from this phase-plane plot
}
\examples{
library(fda)
data(CanadianWeather)
goodsbasis <- create.bspline.basis(rangeval=c(1919,2000), nbasis=161, norder=8)
LfdobjNonDur <- int2Lfd(4)
argvals = seq(1919,2000,len=length(nondurables))
logNondurSm <- smooth.basisPar(argvals, y=log10(nondurables),
fdobj=goodsbasis, Lfdobj=LfdobjNonDur, lambda=1e-11)

log_fd = logNondurSm$fd
XaxisYaxis_span <- xyspan(c(1920,2000),  log_fd)
}
